implement card component so that the layout doesnot gets re-rendered on animation
