+++ src/pages/Dashboard/components/boardSchema.ts
@@ -1,0 +1,7 @@
+export interface BoardSchema {
+  title: string;
+  id: string;
+  owners: string[];
+  members: string[];
+  background: { [prop: string]: string };
+}
