+++ src/pages/Dashboard/components/index.ts
@@ -1,0 +1,2 @@
+export type { BoardSchema } from './boardSchema';
+export { default as DashboardCard } from './DashboardCard';
